--- src/main/java/com/jiho/anniehands/auth/AuthenticationController.java
@@ -1,15 +1,0 @@
-package com.jiho.anniehands.auth;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequiredArgsConstructor
-@RequestMapping("/auth")
-@Slf4j
-public class AuthenticationController {
-
-
-}
