+++ src/main/java/com/jiho/anniehands/common/exception/CustomErrorCode.java
@@ -1,4 +1,4 @@
-package com.jiho.anniehands.exception;
+package com.jiho.anniehands.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
