+++ src/main/java/com/jiho/anniehands/user/UserException.java
@@ -1,7 +1,7 @@
 package com.jiho.anniehands.user;
 
-import com.jiho.anniehands.exception.AnnieHandsException;
-import com.jiho.anniehands.exception.CustomErrorCode;
+import com.jiho.anniehands.common.exception.AnnieHandsException;
+import com.jiho.anniehands.common.exception.CustomErrorCode;
 
 public class UserException extends AnnieHandsException {
     public UserException(CustomErrorCode errorCode) {
