+++ src/main/java/com/jiho/anniehands/user/UserService.java
@@ -1,6 +1,6 @@
 package com.jiho.anniehands.user;
 
-import com.jiho.anniehands.exception.CustomErrorCode;
+import com.jiho.anniehands.common.exception.CustomErrorCode;
 import com.jiho.anniehands.security.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
